refactor: extract connection ID issuer
The ConnectionIdIssuer is responsible for generating and verifing
connection ids.
