+++ src/udp/handlers.rs
@@ -4,9 +4,9 @@ use std::sync::Arc;
 use aquatic_udp_protocol::{AnnounceInterval, AnnounceRequest, AnnounceResponse, ConnectRequest, ConnectResponse, ErrorResponse, NumberOfDownloads, NumberOfPeers, Port, Request, Response, ResponsePeer, ScrapeRequest, ScrapeResponse, TorrentScrapeStatistics, TransactionId, ConnectionId};
 use log::debug;
 
-use crate::udp::connection::cypher::BlowfishCypher;
+use crate::udp::connection::connection_id_issuer::{EncryptedConnectionIdIssuer, ConnectionIdIssuer};
+
 use crate::udp::connection::secret::Secret;
-use crate::udp::connection::connection_id::new_connection_id;
 use crate::{InfoHash, MAX_SCRAPE_TORRENTS};
 use crate::peer::TorrentPeer;
 use crate::tracker::torrent::{TorrentError};
@@ -94,12 +94,13 @@ pub fn generate_new_connection_id(remote_addr: &SocketAddr) -> ConnectionId {
     // todo: server_secret should be randomly generated on startup
     let server_secret = Secret::new([0;32]);
 
-    // todo: this is expensive. It should be generated on startup
-    let cypher = BlowfishCypher::new(server_secret);
+    // todo: this is expensive because of the blowfish cypher instantiation.
+    // It should be generated on startup.
+    let issuer = EncryptedConnectionIdIssuer::new(server_secret);
 
     let current_timestamp = current_timestamp();
 
-    let connection_id = new_connection_id(&cypher, remote_addr, current_timestamp);
+    let connection_id = issuer.new_connection_id(remote_addr, current_timestamp);
 
     debug!("new connection id: {:?}, current timestamp: {:?}", connection_id, current_timestamp);
